+++ .github/workflows/build-image.yaml
@@ -71,7 +71,7 @@ jobs:
             "$IMAGE_TAG" \
             "$GITHUB_SERVER_URL/$GITHUB_REPOSITORY" \
             "$GITHUB_SERVER_URL/$GITHUB_REPOSITORY/actions/runs/$GITHUB_RUN_ID" > ./version.json
-          docker build --file infra/Dockerfile -t $GITHUB_REPOSITORY:$IMAGE_TAG .
+          docker build --file Dockerfile -t $GITHUB_REPOSITORY:$IMAGE_TAG .
           docker image tag $GITHUB_REPOSITORY:$IMAGE_TAG $GITHUB_REPOSITORY:latest
           docker push $GITHUB_REPOSITORY:$IMAGE_TAG
           docker push $GITHUB_REPOSITORY:latest
