+++ .github/workflows/test-build.yaml
@@ -20,4 +20,4 @@ jobs:
       - name: Install dependencies
         run: poetry install
       - name: Run tests
-        run: infra/test.sh
+        run: bin/test.sh
